Add subintent transaction shape tests
